refactor: 중요도 옵션 상수로 리팩토링
